implementacija/stream/producer/requirements.txt
kafka-python
websocket-client
pandas